requirements.txt
# Core framework
Django>=5.0,<7.0

# Data & ML
pandas>=2.0
numpy>=1.26
scikit-learn>=1.3
joblib>=1.3
python-dateutil>=2.8

# Data seeding (dev/demo only)
Faker>=24.0

# Optional: for real XGBoost model (falls back to GradientBoosting if absent)
# xgboost>=2.0

# Production extras (uncomment when deploying)
# gunicorn>=21.0
# whitenoise>=6.6
# psycopg2-binary>=2.9   # if using PostgreSQL
